refactor: put types in a separate file
